Replace opening "<" in element completions
Slightly hacky mechanism for handling behavioural changes in the VSCode extension or LSP API where reported position no longer includes character that was typed to trigger completion (it's possible that this extension is not reporting some newly-defined capability or setting to VSCode that would enable it to see the previous behaviour but it's hard to tell given how often things seem to change).

NOTE: This set of changes is not exhaustive - it's just a record of the types of changes that need to be made more extensively.

tintoy/msbuild-project-tools-vscode#67