Improve language server's detection of target .NET SDK version
We do so by making a distinction between runtime (host) and SDK versions, using the runtime version as the target for the extension / language server itself, and using the SDK version to find an appropriate version of MSBuild.

We will probably need to move the MSBuild stuff out to a separate process at some point in order to ensure that we can target multiple versions of MSBuild, etc.

tintoy/msbuild-project-tools-vscode#73